requirements.txt
django==4.2.2
djangorestframework==3.14.0
django-debug-toolbar==4.2.0